+++ src/components/pages/Project/Project.tsx
@@ -1,13 +1,13 @@
+import { Banner } from "@/components/common/Banner/Banner";
 import { Header } from "@/components/common/Header";
 import { ProjectSelectTab } from "@/components/common/Tabs";
 import { TabType } from "@/components/common/Tabs/ProjectSelectTab/ProjectSelectTab";
 import { useReducer } from "react";
-import { ProjectTabAll } from "./elements/ProjectTabAll";
-import { ProjectTabEvent } from "./elements/ProjectTabEvent";
-import { filterReducer } from "./filterReducer";
+import { ProjectTab } from "./elements/ProjectTab";
 import styles from "./Project.module.css";
-import { Banner } from "@/components/common/Banner/Banner";
-import { BannerList } from "./BannerList";
+import { BannerList } from "./utils/BannerList";
+import { filterReducer } from "./utils/filterReducer";
+import { projectEventProps, projectAllProps } from "./utils/getProjectList";
 
 export function Project() {
   const PROJECT_BANNER_INFO = BannerList.find((item) => item.type === "PROJECT")!;
@@ -22,12 +22,12 @@ export function Project() {
     {
       id: "1",
       label: "S-TOP 이벤트 프로젝트",
-      children: <ProjectTabEvent {...reducerProps} />,
+      children: <ProjectTab {...reducerProps} data={projectEventProps} />,
     },
     {
       id: "2",
       label: "전체 프로젝트",
-      children: <ProjectTabAll {...reducerProps} />,
+      children: <ProjectTab {...reducerProps} data={projectAllProps} />,
     },
   ];
